Добавлен переход в списки при удалении места/города
Resolve https://github.com/stas12312/my-tickets-bot/issues/130
