+++ my_tickets_bot/src/bot/handlers/cities/__init__.py
@@ -1,9 +1,9 @@
 """Обработчики для городов"""
 from aiogram import Router
 
-from . import add, edit, show
+from . import add, delete, show
 
 router = Router()
 router.include_router(add.router)
-router.include_router(edit.router)
+router.include_router(delete.router)
 router.include_router(show.router)
