+++ my_tickets_bot/src/bot/handlers/cities/delete.py
@@ -1,6 +1,7 @@
 from aiogram import types, Router, F
 
 from bot.callbacks import CityCallback, EntityAction
+from bot.services.cities.messages import get_show_cities_params
 from services.repositories import Repo
 
 
@@ -13,7 +14,10 @@ async def delete_city(
 
     city_id = callback_data.city_id
     await repo.city.delete(query.from_user.id, city_id)
-    await query.message.edit_text('Город удалён')
+    msg, keyboard = await get_show_cities_params(query.from_user.id, repo)
+    # Переходим в список городов
+    await query.message.edit_text(msg, reply_markup=keyboard)
+    await query.answer('Город удален')
 
 
 router = Router()
