+++ my_tickets_bot/src/bot/handlers/cities/show.py
@@ -1,32 +1,26 @@
 from aiogram import types, Router, F
 
 from bot.callbacks import CityCallback, EntityAction
-from bot.keybaords import get_cities_menu, get_actions_for_city
-from bot.messages import make_city_message
+from bot.services.cities.keyboards import get_actions_for_city
+from bot.services.cities.messages import make_city_message, get_show_cities_params
 from services.repositories import Repo
 
 
-async def show_cities_handler(
+async def show_cities(
         query: types.CallbackQuery,
         repo: Repo,
 ):
     """Отображение списка городов в inline-клавиатуре"""
-
-    cities = await repo.city.list(query.from_user.id)
-    menu = get_cities_menu(cities)
-    await query.message.edit_text(
-        '🏙 _Ваши города_\n\n'
-        'ℹ️ Выберите город',
-        reply_markup=menu)
+    msg, keyboard = await get_show_cities_params(query.from_user.id, repo)
+    await query.message.edit_text(msg, reply_markup=keyboard)
 
 
-async def show_city_handler(
+async def show_city(
         query: types.CallbackQuery,
         callback_data: CityCallback,
         repo: Repo,
 ):
     """Отображение города"""
-
     city_id = callback_data.city_id
     city = await repo.city.get(query.from_user.id, city_id)
 
@@ -38,5 +32,5 @@ async def show_city_handler(
 
 
 router = Router()
-router.callback_query.register(show_cities_handler, CityCallback.filter(F.action == EntityAction.LIST))
-router.callback_query.register(show_city_handler, CityCallback.filter(F.action == EntityAction.SHOW))
+router.callback_query.register(show_cities, CityCallback.filter(F.action == EntityAction.LIST))
+router.callback_query.register(show_city, CityCallback.filter(F.action == EntityAction.SHOW))
