+++ my_tickets_bot/src/bot/handlers/events/add.py
@@ -7,8 +7,10 @@ from aiogram.types import ContentType
 
 from bot.buttons import Action, MainMenu
 from bot.forms import EventForm
-from bot.keybaords import get_keyboard_by_values, get_add_city_keyboard, get_menu_keyboard
-from bot.messages import get_address, TIME_EXAMPLES
+from bot.keybaords import get_keyboard_by_values, get_menu_keyboard
+from bot.services.cities.keyboards import get_add_city_keyboard
+from bot.messages import TIME_EXAMPLES
+from bot.services.locations.messages import get_address
 from bot.utils import save_ticket
 from services.event_time import parse_datetime, get_localtime
 from services.repositories import Repo
