+++ my_tickets_bot/src/bot/handlers/locations/delete.py
@@ -1,6 +1,7 @@
 from aiogram import types, Router, F
 
 from bot.callbacks import LocationCallback, EntityAction
+from bot.services.locations.messages import get_show_locations_params
 from services.repositories import Repo
 
 
@@ -11,9 +12,12 @@ async def delete_location_handler(
 ):
     """Удаление места"""
     location_id = callback_data.location_id
+    city_id = callback_data.city_id
     await repo.location.delete(query.from_user.id, location_id)
+    msg, keyboard = await get_show_locations_params(query.from_user.id, city_id, repo)
+
+    await query.message.edit_text(msg, reply_markup=keyboard)
     await query.answer('Место удалено')
-    await query.message.delete()
 
 
 router = Router()
