+++ my_tickets_bot/src/bot/keybaords.py
@@ -9,7 +9,7 @@ from aiogram.utils.keyboard import (
     InlineKeyboardBuilder,
 )
 
-from models import City, Location, Event, Ticket
+from models import Event, Ticket
 from .buttons import (
     MainMenu,
     Settings,
@@ -20,8 +20,6 @@ from .buttons import (
 from .callbacks import (
     CityCallback,
     EntityAction,
-    LocationCallback,
-    SettingsCallback,
     CloseCallback,
     TicketCallback,
     EventCallback,
@@ -70,123 +68,6 @@ def get_settings_menu() -> InlineKeyboardMarkup:
         ),
     )
     builder.row(CLOSE_BUTTON)
-
-    return builder.as_markup()
-
-
-def get_cities_menu(
-        cities: list[City]
-) -> InlineKeyboardMarkup:
-    """Получение клавиатуры для выбора города"""
-    builder = InlineKeyboardBuilder()
-
-    for city in cities:
-        builder.row(
-            InlineKeyboardButton(
-                text=city.name,
-                callback_data=CityCallback(action=EntityAction.SHOW, city_id=city.city_id).pack(),
-            )
-        )
-
-    builder.row(
-        InlineKeyboardButton(
-            text=Settings.ADD_CITY,
-            callback_data=CityCallback(action=EntityAction.ADD).pack(),
-        )
-    )
-    builder.row(
-        InlineKeyboardButton(
-            text=Settings.BACK,
-            callback_data=SettingsCallback(action=EntityAction.SHOW).pack(),
-        ),
-        CLOSE_BUTTON,
-    )
-
-    return builder.as_markup()
-
-
-def get_actions_for_city(
-        city_id: int,
-) -> InlineKeyboardMarkup:
-    """Получение клавиатуры с действиями для города"""
-    builder = InlineKeyboardBuilder()
-
-    builder.row(
-        InlineKeyboardButton(
-            text=Settings.PLACES,
-            callback_data=LocationCallback(action=EntityAction.LIST, city_id=city_id).pack()
-        )
-    )
-
-    builder.row(
-        InlineKeyboardButton(
-            text=Settings.DELETE_CITY,
-            callback_data=CityCallback(action=EntityAction.DELETE, city_id=city_id).pack(),
-        )
-    )
-    builder.row(
-        InlineKeyboardButton(
-            text=Settings.BACK,
-            callback_data=CityCallback(action=EntityAction.LIST).pack(),
-        ),
-        CLOSE_BUTTON,
-    )
-
-    return builder.as_markup()
-
-
-def get_locations_menu(
-        city_id: int,
-        locations: list[Location],
-) -> InlineKeyboardMarkup:
-    """Получение меню для списка мест"""
-    builder = InlineKeyboardBuilder()
-
-    for location in locations:
-        builder.row(
-            InlineKeyboardButton(
-                text=location.name,
-                callback_data=LocationCallback(action=EntityAction.SHOW, location_id=location.location_id).pack(),
-            )
-        )
-
-    builder.row(
-        InlineKeyboardButton(
-            text=Settings.ADD_LOCATION,
-            callback_data=LocationCallback(action=EntityAction.ADD, city_id=city_id).pack(),
-        )
-    )
-    builder.row(
-        InlineKeyboardButton(
-            text=Settings.BACK,
-            callback_data=CityCallback(action=EntityAction.SHOW, city_id=city_id).pack(),
-        ),
-        CLOSE_BUTTON,
-    )
-
-    return builder.as_markup()
-
-
-def get_actions_for_location(
-        city_id: int,
-        location_id: int,
-) -> InlineKeyboardMarkup:
-    """Получение клавиатуры для действия с местом"""
-    builder = InlineKeyboardBuilder()
-
-    builder.row(
-        InlineKeyboardButton(
-            text=Settings.DELETE_LOCATION,
-            callback_data=LocationCallback(action=EntityAction.DELETE, location_id=location_id).pack(),
-        )
-    )
-    builder.row(
-        InlineKeyboardButton(
-            text=Settings.BACK,
-            callback_data=LocationCallback(action=EntityAction.LIST, city_id=city_id).pack(),
-        ),
-        CLOSE_BUTTON,
-    )
 
     return builder.as_markup()
 
@@ -245,20 +126,6 @@ def get_actions_for_ticket(
     )
 
     return builder.as_markup()
-
-
-def get_add_city_keyboard() -> InlineKeyboardMarkup:
-    """Клавиатура для добавления города"""
-    return InlineKeyboardBuilder([
-        [
-            InlineKeyboardButton(
-                text=Settings.ADD_CITY,
-                callback_data=CityCallback(
-                    action=EntityAction.ADD,
-                ).pack()
-            ),
-        ],
-    ]).as_markup()
 
 
 def get_actions_for_edit_event(
